+++ ports/servoshell/desktop/protocols/mod.rs
@@ -5,3 +5,4 @@
 pub(crate) mod resource;
 pub(crate) mod servo;
 pub(crate) mod urlinfo;
+pub(crate) mod view_source;